Isolate wells without preprocessing images
The Single Worm Output is simplified to a point where it identifies
wells, draw the grid and isolate wells as individual images. This is
a test version to which negated and thresholded input image has to be
given manually. This version does not work with the script yet and
output cannot be processed by the RandIndex pipeline.