Use mask and change movement quantification method
The script generates a mask from the first frame which is applied to all
subsequent frames. This ensures that the exact same grid is drawn on
each frame and that the image of an isolated well will have always the
same size which is important for the new movement quantification method.

Quantification of movement is now based on computing the image difference
(error) between the same well overtime. This is performed by the ImageMagick
package. The consequences of this are:
* The difference is computed between each frame which was not the case
with the RandIndex. The RandIndex compared image 1-2 and 3-5 but not 2-3
which left almost a third to half of the data unused.
* The RandIndex pipeline is not needed anymore and has been removed.