Add a new script to determine the best grid
Determining the best setting to draw the grid at the expected position
on the plate requires to fiddle with the threshold value. However,
running the full pipeline for each test is time consumming. The best
grid script helps to reduce the time spent by running a truncated version
of the pipeline.