Analyze and summarize movement from each well
Add a new step in the main script to analyze the table generated after
CellProfiler processing. This step relies on a R script which produces a
final table of average values of movement per well and useful graphs.
A new feature regarding well numbering and a progress bar to monitor
CellProfiler processing have been added to the main script.