Add a test mode option to the pipeline
Threshold parameter to optimally place the grid for well detection should
be determined using the best-grid.sh script. However, running the
pipeline with the determined threshold often led to unexpected results.
This is due to the differences in image processing between the two
scripts. To avoid this, a step mode has been added to the main script
to run the pipeline on a single frame. This led to the removal of the
best-grid.sh script and its dependency.
In addition, detection of wells has been improved by changing the fuzz
option. A new step ensure that the requested stop time is compatible
with the length of video has been added.